server: shuffle platform dependent operations
