+++ server/label_unsupported.go
@@ -1,0 +1,7 @@
+// +build !linux
+
+package server
+
+func securityLabel(path string, seclabel string, shared bool) error {
+	return nil
+}
