+++ server/sandbox_run_unsupported.go
@@ -1,0 +1,14 @@
+// +build !linux
+
+package server
+
+import (
+	"fmt"
+
+	"golang.org/x/net/context"
+	pb "k8s.io/kubernetes/pkg/kubelet/apis/cri/runtime/v1alpha2"
+)
+
+func (s *Server) runPodSandbox(ctx context.Context, req *pb.RunPodSandboxRequest) (resp *pb.RunPodSandboxResponse, err error) {
+	return nil, fmt.Errorf("unsupported")
+}
